Intelligent Background Feature
Intelligent Background is a feature that colors the 4 corners of a card when printing completely square cards instead of the original round corners.
Instead of filling the corners black, it adapts the color the each corner individually.
This is especially useful for borderless cards since these often come with non-black corners.